Added in Settings Page
Added in the frontend to the settings page, and added the connection to the backend through the creation of an API blueprint. It first recieves the current user data (Only a placeholder right now) and then updates the current page based on that, then it has an event listener looking for a setting change and then sends it back to the server.